x86-64: implement movabs (mov r64, imm64)
mov r,imm (B8+r) is the one instruction whose immediate scales to the full
operand size: imm16 under 66, imm32 by default, and imm64 under REX.W
(movabs). It was wired to IMK_IMMZ, which caps at 32 bits, so a REX.W
B8+r decoded only 4 of its 8 immediate bytes -- the remaining 4 were then
mis-decoded as a bogus following instruction. gcc emits movabs for large
constants and absolute addresses, so this broke round-trip on real code.

Add IMK_IMMV (operand-size immediate, imm64 under REX.W) and point the
B8+r mov rule at a new `immv` submatch. REX.W is not an FSM var (it is
consumed raw into pfx[]), so the width is resolved C++-side: decode_insn
derives rexw from the effective last prefix and passes it to append_imm
(IMK_IMMV -> 2/4/8 bytes by 66/default/REX.W); enc_imm mirrors it from
in->opsize. The immediate field widens from int32_t to int64_t so the
64-bit value fits, and finalize_insn keeps the full 64 bits for the
movabs case (narrower immediates still occupy the low bytes, and the
encoder writes only as many bytes as the operand size needs). gen.py and
gasm.py classify `immv` via the existing name->IMK_* path; no encoder
candidate change is needed (movabs stays FORM_REG_IMM, REX.B-extended).

The int64_t imm widening is inert for x86-32: it never produces a 64-bit
immediate and the encoder writes only the low bytes, so the byte stream
is unchanged. Added 4 movabs + 3 mov-imm-width variants to corpus64.bin
for permanent regression coverage.

Validation:
  x86-32 roundtrip      858/858 byte-identical
  x86-32 x8632all       110072/110072 byte-identical (18836 insns)
  x86-32 fuzz 5M        0 legacy, 0 vex failures
  x86-64 roundtrip64    byte-identical (48 insns, now incl. movabs)
  x86-64 prog64         byte-identical (349 gcc -O2 integer insns)
  x86-64 fuzz64 5M      0 legacy failures (VEX/EVEX/XOP deferred)
  movabs unit           mov r64,imm64 / r15 / rcx,-1 / r8 round-trip exact

+++ fsm-port/012/x86insn.hpp
@@ -71,8 +71,10 @@ typedef struct {
   uint8_t enc    : 3;  /* opcode-choice deviation: 0=canonical, else dir/long-imm/acc/moffs   */
   uint8_t reg_w  : 4;  /* dead ModR/M reg field for single-r/m forms (setcc), REX.R-extended  */
 
-  /* the single full immediate (target of T_IMM / T_REL / T_PTR) */
-  int32_t imm;
+  /* the single full immediate (target of T_IMM / T_REL / T_PTR). 64-bit so the
+   * REX.W mov r64,imm64 (movabs) immediate fits; narrower immediates occupy the
+   * low bytes (the encoder writes only as many bytes as the operand size needs). */
+  int64_t imm;
 
   uint8_t cc;          /* condition code 0..15 for jcc/setcc/cmovcc; 0xFF = none */
